+++ totmann.inc.dist.php
@@ -213,8 +213,7 @@ This is a reminder to confirm that you are safe and able to respond.
 Please click this link to confirm:
 {CONFIRM_URL}
 
-You must click the link no later than the confirmation deadline!
-
+You must click the link no later than the confirmation deadline.
 Confirmation deadline: {DEADLINE_ISO}
 Current cycle started: {CYCLE_START_ISO}